requirements.txt
albumentations>=1.1.0
aws-cdk-lib>=2.14.0
boto3>=1.21.8

hydra-core>=1.2.0
pytest>=7.0.1
tox>=3.24.5
black>=22.1.0
isort>=5.10.1
mypy>=0.941
flake8>=4.0.1
pydantic>=1.8.2

ray>=1.10.0

scikit-image>=0.19.2
scikit-learn>=1.0.2

tensorboard>=2.8.0
transformers>=4.15.0

tqdm>=4.62.3
python-dotenv>=0.19.2
matplotlib>=3.5.1
japanize-matplotlib>=1.1.3

jupyter>=1.0.0
jupyterlab>=3.2.9
pandas>=1.4.1
joblib>=1.1.0

matplotlib>=1.5
numpy>=1.22.2
scipy>=1.8.0
seaborn>=0.11.2
pytorch-lightning>=1.5.10
opencv-contrib-python>=4.5.5.62
opencv-python>=4.5.5.62
Pillow>=9.0.1

datasets>=1.18.3
fugashi>=1.1.2
SudachiPy>=0.6.3
sudachidict_core
ja-ginza>=5.1.0
jaconv>=0.3
ginza>=5.1.0
ginza-transformers>=0.4.0

spacy>=3.2.2
torchmetrics>=0.7.2
nnmnkwii>=0.1.1
pyopenjtalk>=0.1.0

neptune-client==0.14.3
dvc==2.9.5

nptyping==2.2.0